+++ src/Symfony/Component/Serializer/CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 8.2
 ---
 
+ * Add `BsonEncoder` and `BsonDecoder` to encode and decode MongoDB BSON documents and arrays, requires the `mongodb` extension
  * Add the `force_collection` context option to `XmlEncoder` to always decode the given tags as a collection
  * Add `AbstractNormalizer::SKIP_INVALID_ATTRIBUTES` to denormalize the attributes whose value cannot be used as if they were absent from the input
  * Add `AbstractNormalizer::IGNORED_GROUPS` to exclude the attributes belonging to the given groups